AnonMessageBot/requirements.txt
aiogram>=3.2.0
python-dotenv==1.0.0
redis>=5.0.1 